Mostly style/whitespace, a few substantive changes
Replaced \t with "  " everywhere, and shortened lines to less than 80
chars.

Cleaned up the token.registry.js a lot, removing some unnecessary code,
and set it up to clean out every time a token is created, if there are a
lot of tokens already.

Fixed a few actual bugs in couch.facade.js and server.js that were
apparent once the indentation was corrected.

Moved all the configuration params to environs, so there's no need to
read through the code and replace hard-coded email info.